+++ demos/react-native-supabase-group-chat/src/app/(app)/(chats)/g/[group]/index.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react';
 import { Input, YStack } from 'tamagui';
 
 import { Message } from '@/components/messages/Message';
-import { uuid } from '@/lib/uuid';
+import { uuid } from '@/library/uuid';
 import { useAuth } from '@/providers/AuthProvider';
 
 export default function ChatsChatIndex() {
@@ -44,7 +44,7 @@ export default function ChatsChatIndex() {
       />
       <YStack fullscreen>
         <YStack flexGrow={1}>
-          <FlashList data={messages} renderItem={({ item }) => <Message message={item} />} estimatedItemSize={87} />
+          <FlashList data={messages} renderItem={({ item }) => <Message message={item} />} />
         </YStack>
         <YStack padding="$3" gap="$3">
           <Input